# AppBackend/openapi.yaml
openapi: 3.0.3
info:
  title: PhotoApp Backend API
  version: 1.0.0
  description: OpenAPI specification for PhotoApp backend (auto-generated from controller files)
servers:
  - url: http://localhost:3000
...
paths:
...
  /photos/upload:
    post:
      tags:
        - Photos
      summary: Upload multiple photos (auth required)
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photos:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Upload result with photos list
  /photos:
    get:
      tags:
        - Photos
      summary: List current user's photos (auth required)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of photos
  /photos/delete/{photoName}:
    post:
      tags:
        - Photos
      summary: Delete photo by name (auth required)
      security:
        - bearerAuth: []
      parameters:
        - name: photoName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
...
  /photos/delete-by-url:
    post:
      tags:
        - Photos
      summary: Delete photo by URL (auth required)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                photoUrl:
                  type: string
      responses:
        '200':
          description: Deleted
  /photos/library:
    get:
      tags:
        - Photos
      summary: List library photos grouped by year (auth required)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Grouped photos
  /photos/batch-delete:
    post:
      tags:
        - Photos
      summary: Batch delete photos by names (auth required)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                photoNames:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Batch delete result
  /photos/connect-drive:
    post:
      tags:
        - Photos
      summary: Generate Google Drive auth URL (auth required)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Google Drive auth URL generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  authUrl:
                    type: string
  /photos/drive-callback:
    post:
      tags:
        - Photos
      summary: Exchange code for tokens and save (auth required)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
      responses:
        '200':
          description: Drive connected
  /photos/drive-status:
    get:
      tags:
        - Photos
      summary: Check Google Drive connection status (auth required)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Google Drive connection status
          content:
            application/json:
              schema:
                type: object
                properties:
                  connected:
                    type: boolean
                  valid:
                    type: boolean
  /photos/disconnect-drive:
    post:
      tags:
        - Photos
      summary: Disconnect Google Drive (auth required)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Disconnected
...
  /albums/{id}/photos:
    post:
      tags:
        - Albums
      summary: Add photo to album (auth required)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                photoName:
                  type: string
                photoUrl:
                  type: string
      responses:
        '200':
          description: Photo added
    get:
      tags:
        - Albums
      summary: Get photos from album (auth required)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Album photos
...
  /categories:
    post:
      tags:
        - Categories
      summary: Create category (auth required)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                color:
                  type: string
      responses:
        '201':
          description: Category created
    get:
      tags:
        - Categories
      summary: Get user categories (auth required)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Categories list
  /categories/{id}:
    get:
      tags:
        - Categories
      summary: Get category by id (auth required)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Category
    put:
      tags:
        - Categories
      summary: Update category (auth required)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                color:
                  type: string
      responses:
        '200':
          description: Updated category
  /categories/{id}/delete:
    post:
      tags:
        - Categories
      summary: Delete category (auth required)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT